+++ helpers/__tests__/teamspace.test.js
@@ -15,7 +15,7 @@ beforeAll(() => {
   global.DataStore = DataStore
   global.Editor = Editor
   global.NotePlan = new NotePlan()
-  DataStore.settings['_logLevel'] = 'none' // change this to DEBUG to get more logging, or 'none' for quiet
+  DataStore.settings['_logLevel'] = 'DEBUG' // change this to DEBUG to get more logging, or 'none' for quiet
 })
 
 const PLUGIN_NAME = `📙 ${colors.yellow('helpers/teamspace')}`
@@ -31,7 +31,7 @@ describe(`${PLUGIN_NAME}`, () => {
       expect(result).toEqual({ filename: '20250422.md', isTeamspace: false })
     })
     test('should parse a Teamspace calendar filename', () => {
-      const result = t.parseTeamspaceFilename('%%Supabase%%/c484b190-77dd-4d40-a05c-e7d7144f24e1/20250422.md')
+      const result = t.parseTeamspaceFilename('%%NotePlanCloud%%/c484b190-77dd-4d40-a05c-e7d7144f24e1/20250422.md')
       expect(result).toEqual({ filename: '20250422.md', isTeamspace: true, teamspaceID: 'c484b190-77dd-4d40-a05c-e7d7144f24e1' })
     })
   })